// ClientApp/src/store/post/post.types.ts
import { Comment } from "../comment/comment.types";
import { User } from "../user/user.types";

export enum POST_ACTION_TYPES {
    CREATE_START = 'post/CREATE_START',
    CREATE_SUCCESS = 'post/CREATE_SUCCESS',
    CREATE_FAILED = 'post/CREATE_FAILED',
    UPDATE_START = 'post/UPDATE_START',
    UPDATE_SUCCESS = 'post/UPDATE_SUCCESS',
    UPDATE_FAILED = 'post/UPDATE_FAILED',
    DELETE_START = 'post/DELETE_START',
    DELETE_SUCCESS = 'post/DELETE_SUCCESS',
    DELETE_FAILED = 'post/DELETE_FAILED',
    FETCH_SINGLE_START = 'post/FETCH_SINGLE_START',
    FETCH_SINGLE_SUCCESS = 'post/FETCH_SINGLE_SUCCESS',
    FETCH_SINGLE_FAILED = 'post/FETCH_SINGLE_FAILED',
    FETCH_ALL_START = 'post/FETCH_ALL_START',
    FETCH_ALL_SUCCESS = 'post/FETCH_ALL_SUCCESS',
    FETCH_ALL_FAILED = 'post/FETCH_ALL_FAILED'
};

export type Post = {
    postId: number | null;
    postValue: string;
    mediaLink: string;
    imageFile: File | null;
    dateCreated: Date;
    userId: number;
    user: User;
    comments: Comment[];
}